requirements.txt
# Principais
fastapi==0.127.1
uvicorn==0.40.0
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.4